+++ java-notebooks/google-cloud-notebooks/src/test/java/com/google/cloud/notebooks/v1beta1/it/Util.java
@@ -22,7 +22,8 @@ public class Util {
         ListInstancesRequest.newBuilder().setParent(parent).build());
     for (Instance instance : listInstancesPagedResponse.iterateAll()) {
       if (isCreatedBeforeThresholdTime(
-          Instant.ofEpochMilli(Timestamps.toMillis(instance.getCreateTime())))) {
+          Instant.ofEpochMilli(Timestamps.toMillis(instance.getCreateTime()))) &&
+          instance.getName().startsWith(ITNotebookServiceClientTest.NOTEBOOK_PREFIX)) {
         client.deleteInstanceAsync(
             DeleteInstanceRequest.newBuilder().setName(instance.getName()).build());
       }
